Inventory Managment System Project Java
This is a project on inventory mangment system in which the scenario is we purchase bulk product from companies and then sale them individually with some increased price. The product purchased in added the stock on every sale the product quantity is subtracted from the stock and then sale is then transfered to the invoice and a slip can be generated.